Merge

+++ JobPortal.IRepository/IRepository.cs
@@ -9,7 +9,7 @@ namespace JobPortal.IRepository
     public interface IRepository<T> where T : class
     {
         Task<IEnumerable<T>> GetAllAsync();
-        Task<T> GetAsync(int id);
+        Task<T> GetAsync(long id);
         Task<T> CreateAsync(T entity);
         Task<T> UpdateAsync(T entity);
         Task<bool> DeleteAsync(T entity);
